fix: design-system docs build
