+++ packages/design-system/src/components/OcRecipient/types.ts
@@ -1,0 +1,10 @@
+export type Recipient = {
+  name: string
+  icon?: {
+    name: string
+    label: string
+  }
+  isLoadingAvatar?: boolean
+  hasAvatar?: boolean
+  avatar?: string
+}
