+++ packages/web-app-files/src/components/SideBar/Shares/Collaborators/InviteCollaborator/RecipientContainer.vue
@@ -21,7 +21,7 @@
 import { avatarUrl } from '../../../../../helpers/user'
 import { ShareTypes } from '@ownclouders/web-client/src/helpers/share'
 import { defineComponent } from 'vue'
-import { Recipient } from 'design-system/src/components/OcRecipient/OcRecipient.vue'
+import { Recipient } from 'design-system/src/components/OcRecipient/types'
 import { useCapabilityStore, useConfigStore } from '@ownclouders/web-pkg'
 import { storeToRefs } from 'pinia'
